scraper/requirements.txt
psycopg2-binary>=2.9
requests>=2.28
feedparser>=6.0
python-dotenv>=1.0
google-api-python-client>=2.100
google-generativeai>=0.3.0
langdetect
isodate
